user_service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["naveen-ravi2810 <rnaveen28102003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
alembic = "^1.13.1"
uuid7 = "^0.1.0"
ruff = "^0.3.5"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
